⚡️ feat: 一旦、クラス構成を見直した。
この時点では、lオプションは正しく動作しないが、含めてcommitしている。
