+++ 07.ls_object/format.rb
@@ -1,77 +1,58 @@
 # frozen_string_literal: true
 
-require_relative 'ls'
+require_relative 'myfilestat'
 
 class Format
-  def initialize(opts, file_list)
-    @format_options = opts
+  INITIAL_COLUMN = 3
+
+  def initialize(file_list)
     @file_list = file_list
   end
 
-  def format
-    if @format_options.empty? || !@format_options.include?('l')
-      Ls.ls_column(@file_list)
+  def ls_column
+    rows, width = calculate_row_and_space
+    if @file_list.size <= 3
+      @file_list.each { |f| print f.ljust(width) }
+      puts
+    elsif @file_list.size == 4
+      transposed_files = @file_list.each_slice(rows).to_a.transpose
+      transposed_files[0].each { |tf| print tf.ljust(width) }
+      print transposed_files[1][0].ljust(width)
+      puts
+      puts transposed_files[1][1]
     else
-      file_list_with_info, total_blocks = ls_opt_long
-      Ls.print_file_list(file_list_with_info, total_blocks)
+      all_sort_files = @file_list.each_slice(rows).to_a
+      rows.times do |col|
+        INITIAL_COLUMN.times do |row|
+          file_name = all_sort_files[row][col]
+          print file_name.ljust(width) unless file_name.nil?
+        end
+        puts
+      end
     end
   end
 
-  private
-
   def ls_opt_long
-    file_list_with_info, total_blocks = compile_file_stat
-    file_list_with_info = file_list_with_info.transpose.each_with_index { |file_attribute, i| adjust_elements(file_attribute, i) }.transpose
+    fs = MyFileStat.new(@file_list)
+    fs.info.transpose.each_with_index { |file_attribute, i| adjust_elements(file_attribute, i) }.transpose
     [file_list_with_info, total_blocks]
   end
 
-  def compile_file_stat
-    total_blocks = 0
-    file_list = @file_list.map do |file|
-      fs = File::Stat.new(file)
-      file_mtime = fs.mtime
-      total_blocks += fs.blocks
-      [
-        convert_to_file_type(fs) + convert_to_file_permission(fs),
-        fs.nlink.to_i,
-        Etc.getpwuid(fs.uid).name,
-        Etc.getgrgid(fs.gid).name,
-        fs.size,
-        file_mtime.month,
-        file_mtime.day,
-        file_mtime.strftime('%H:%M'),
-        file
-      ]
+
+  private
+
+  def calculate_row_and_space
+    div, mod = @file_list.size.divmod(INITIAL_COLUMN)
+    rows = mod.zero? ? div : (div + 1)
+    width = @file_list.max_by(&:length).length + 7
+    [rows, width]
+  end
+
+  def print_file_list(file_list, total_blocks)
+    puts "total #{total_blocks}"
+    file_list.each do |list|
+      puts list.join(' ')
     end
-    [file_list, total_blocks]
-  end
-
-  def convert_to_file_type(file_stat)
-    {
-      file: '-',
-      directory: 'd',
-      characterSpecial: 'c',
-      blockSpecial: 'b',
-      fifo: 'p',
-      link: 'l',
-      socket: 's',
-      unknown: 'u'
-    }[file_stat.ftype.to_sym]
-  end
-
-  def convert_to_file_permission(file_stat)
-    file_permission_number = file_stat.mode.to_s(8).slice(-3..-1)
-    file_permission_string = {
-      0 => '---',
-      1 => '--x',
-      2 => '-w-',
-      3 => '-wx',
-      4 => 'r--',
-      5 => 'r-x',
-      6 => 'rw-',
-      7 => 'rwx'
-    }
-    file_permission_number.each_char.map { |c| file_permission_string[c.to_i] }.join
   end
 
   def adjust_elements(file_attribute, idx)
