+++ 07.ls_object/main.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'filelist'
-require_relative 'format'
+require_relative 'ls'
 
-option = Option.new.extract_entered_option
-fl = FileList.new(option)
-Format.new(option, fl.files).format
+Ls.new.show
